Customizing style and adding sqlite
New buttons to erase (not implemented) and save paintings, dialogs for
that purpose.
Fix lines were added more than once to a painting, it was nonsense and
also a leak in performance.
Added sqlite to the App by means of a content provider. Paintings are
being stored.
Shared Preference utils to manage autosaving of a painting if the app
is closed.
Service added (not implemented) to store a painting in background.